requirements.txt
aiogram>=3.4,<3.7
aiosqlite>=0.19
openpyxl>=3.1
google-generativeai>=0.5
python-dotenv>=1.0
